Added First step and second
I initially began by using the ULTRA site, in its current state, as a project to learn svelte, so I acted as if I was handed a design file and needed to make the HTML and CSS for that sites index.

From there, that was the template I began with to learn both Svelte and contentful.

Saved that into the first step folder, thinking, someone else may just want to start there.....